DataContainers

EnsembleKalmanProcesses.DataContainers.DataContainer — Type

struct DataContainer{FT<:Real}

Container to store data samples as columns in an array.

Fields

data::AbstractMatrix{FT} where FT<:Real: stored data, each piece of data is a column [data dimension × number samples]

Constructors

DataContainer(data::AVorM; data_are_columns = true) where {AVorM <: AbstractVecOrMat}

EnsembleKalmanProcesses.DataContainers.PairedDataContainer — Type

struct PairedDataContainer{FT<:Real}

Stores input - output pairs as data containers, there must be an equal number of inputs and outputs.

Fields

inputs::EnsembleKalmanProcesses.DataContainers.DataContainer: container for inputs, each Container holds an array size [data/parameter dimension × number samples]

outputs::EnsembleKalmanProcesses.DataContainers.DataContainer: container for ouputs, each Container holds an array size [data/parameter dimension × number samples]

Constructors

PairedDataContainer(
    inputs::AVorM1,
    outputs::AVorM2;
    data_are_columns = true,
) where {AVorM1 <: AbstractVecOrMat, AVorM2 <: AbstractVecOrMat}

PairedDataContainer(inputs::DataContainer, outputs::DataContainer)

Base.size — Function

size(
    c::EnsembleKalmanProcesses.ParameterDistributions.ConstraintType
) -> Tuple{Int64}

A constraint has size 1.

size(dc::DataContainer, idx::IT) where {IT <: Integer}

Returns the size of the stored data. If idx provided, it returns the size along dimension idx.

size(pdc::PairedDataContainer, idx::IT) where {IT <: Integer}

Returns the sizes of the inputs and ouputs along dimension idx (if provided).