EnsembleKalmanProcessModule

EnsembleKalmanProcesses.EnsembleKalmanProcessModule.Inversion — Type

Inversion <: Process

An ensemble Kalman Inversion process

EnsembleKalmanProcesses.EnsembleKalmanProcessModule.Sampler — Type

Sampler{FT<:AbstractFloat,IT<:Int} <: Process

An ensemble Kalman Sampler process

EnsembleKalmanProcesses.EnsembleKalmanProcessModule.EnsembleKalmanProcess — Type

EnsembleKalmanProcess{FT<:AbstractFloat, IT<:Int}

Structure that is used in Ensemble Kalman processes

#Fields

u

Array of stores for parameters (u), each of size [Npar × Nens]

obs_mean

vector of the observed vector size [N_obs]

obs_noise_cov

covariance matrix of the observational noise, of size [Nobs × Nobs]

N_ens

ensemble size

g

Array of stores for forward model outputs, each of size [Nobs × Nens]

err

vector of errors

Δt

vector of timesteps used in each EK iteration

process

the particular EK process (Inversion or Sampler or Unscented)

EnsembleKalmanProcesses.EnsembleKalmanProcessModule.get_u — Function

get_u(ekp::EnsembleKalmanProcess, iteration::IT; return_array=true) where {IT <: Integer}

Get for the EKI iteration. Returns a DataContainer object unless array is specified.

get_u(ekp::EnsembleKalmanProcess; return_array=true)

Get for the EKI iteration. Returns a DataContainer object unless array is specified.

EnsembleKalmanProcesses.EnsembleKalmanProcessModule.get_g — Function

get_g(ekp::EnsembleKalmanProcess, iteration::IT; return_array=true) where {IT <: Integer}

Get for the EKI iteration. Returns a DataContainer object unless array is specified.

get_g(ekp::EnsembleKalmanProcess; return_array=true)

Get for the EKI iteration. Returns a DataContainer object unless array is specified.

EnsembleKalmanProcesses.EnsembleKalmanProcessModule.get_u_final — Function

get_u_final(ekp::EnsembleKalmanProcess, return_array=true)

Get the final or prior iteration of parameters or model ouputs, returns a DataContainer Object if return_array is false.

EnsembleKalmanProcesses.EnsembleKalmanProcessModule.get_N_iterations — Function

get_N_iterations(ekp::EnsembleKalmanProcess

get number of times update has been called (equals size(g), or size(u)-1)

EnsembleKalmanProcesses.EnsembleKalmanProcessModule.construct_initial_ensemble — Function

construct_initial_ensemble(prior::ParameterDistribution, N_ens::IT; rng_seed=42) where {IT<:Int}

Construct the initial parameters, by sampling N_ens samples from specified prior distribution. Returned with parameters as columns

EnsembleKalmanProcesses.EnsembleKalmanProcessModule.find_ekp_stepsize — Function

findekpstepsize(ekp::EnsembleKalmanProcess{FT, IT, Inversion}, g::Array{FT, 2}; cov_threshold::FT=0.01) where {FT}

Find largest stepsize for the EK solver that leads to a reduction of the determinant of the sample covariance matrix no greater than cov_threshold.

EnsembleKalmanProcesses.EnsembleKalmanProcessModule.update_ensemble! — Function

update_ensemble!(ekp::EnsembleKalmanProcess{FT, IT, <:Inversion}, g::Array{FT,2} cov_threshold::FT=0.01, Δt_new=nothing) where {FT, IT}

Updates the ensemble according to which type of Process we have. Model outputs g need to be a Nobs × Nens array (i.e data are columms)