EnsembleKalmanProcesses

EnsembleKalmanProcesses.Inversion — Type

Inversion <: Process

An ensemble Kalman Inversion process

EnsembleKalmanProcesses.Sampler — Type

Sampler{FT<:AbstractFloat,IT<:Int} <: Process

An ensemble Kalman Sampler process.

EnsembleKalmanProcesses.EnsembleKalmanProcess — Type

EnsembleKalmanProcess{FT <: AbstractFloat, IT <: Int, P <: Process}

Structure that is used in Ensemble Kalman processes.

Fields

u::Array{EnsembleKalmanProcesses.DataContainers.DataContainer{FT}} where FT<:AbstractFloat

array of stores for parameters (u), each of size [N_par × N_ens]

obs_mean::Vector{FT} where FT<:AbstractFloat

vector of the observed vector size [N_obs]

obs_noise_cov::Union{LinearAlgebra.UniformScaling{FT}, AbstractMatrix{FT}} where FT<:AbstractFloat

covariance matrix of the observational noise, of size [N_obs × N_obs]

N_ens::Int64

ensemble size

g::Array{EnsembleKalmanProcesses.DataContainers.DataContainer{FT}} where FT<:AbstractFloat

Array of stores for forward model outputs, each of size [N_obs × N_ens]

err::Vector{FT} where FT<:AbstractFloat

vector of errors

Δt::Vector{FT} where FT<:AbstractFloat

vector of timesteps used in each EK iteration

process::EnsembleKalmanProcesses.Process

the particular EK process (Inversion or Sampler or Unscented or SparseInversion)

rng::Random.AbstractRNG

Random number generator object (algorithm + seed) used for sampling and noise, for reproducibility. Defaults to Random.GLOBAL_RNG.

failure_handler::FailureHandler

struct storing failsafe update directives, implemented for (Inversion, SparseInversion, Unscented)

localizer::EnsembleKalmanProcesses.Localizers.Localizer

Localization kernel, implemented for (Inversion, SparseInversion, Unscented)

EnsembleKalmanProcesses.get_u — Function

get_u(ekp::EnsembleKalmanProcess, iteration::IT; return_array=true) where {IT <: Integer}

Get for the EKI iteration. Returns a DataContainer object unless array is specified.

get_u(ekp::EnsembleKalmanProcess; return_array=true)

Get for the EKI iteration. Returns a DataContainer object unless array is specified.

EnsembleKalmanProcesses.get_g — Function

get_g(ekp::EnsembleKalmanProcess, iteration::IT; return_array=true) where {IT <: Integer}

Get for the EKI iteration. Returns a DataContainer object unless array is specified.

get_g(ekp::EnsembleKalmanProcess; return_array=true)

Get for the EKI iteration. Returns a DataContainer object unless array is specified.

EnsembleKalmanProcesses.get_u_final — Function

get_u_final(ekp::EnsembleKalmanProcess, return_array=true)

Get the final or prior iteration of parameters or model ouputs, returns a DataContainer Object if return_array is false.

EnsembleKalmanProcesses.get_N_iterations — Function

get_N_iterations(ekp::EnsembleKalmanProcess

Get number of times update has been called (equals size(g), or size(u)-1).

EnsembleKalmanProcesses.construct_initial_ensemble — Function

construct_initial_ensemble(rng::AbstractRNG, prior::ParameterDistribution, N_ens::IT; rng_seed::Union{IT, Nothing} = nothing)

Construct the initial parameters, by sampling N_ens samples from specified prior distribution. Returned with parameters as columns.

EnsembleKalmanProcesses.find_ekp_stepsize — Function

find_ekp_stepsize(ekp::EnsembleKalmanProcess{FT, IT, Inversion}, g::AbstractMatrix{FT}; cov_threshold::FT=0.01) where {FT}

Find largest stepsize for the EK solver that leads to a reduction of the determinant of the sample covariance matrix no greater than cov_threshold.

EnsembleKalmanProcesses.update_ensemble! — Function

update_ensemble!(
    ekp::EnsembleKalmanProcess{FT, IT, Inversion},
    g::AbstractMatrix{FT};
    cov_threshold::FT = 0.01,
    Δt_new::Union{Nothing, FT} = nothing,
    deterministic_forward_map::Bool = true,
    failed_ens = nothing,
) where {FT, IT}

Updates the ensemble according to an Inversion process.

Inputs:

ekp :: The EnsembleKalmanProcess to update.

g :: Model outputs, they need to be stored as a N_obs × N_ens array (i.e data are columms).

cov_threshold :: Threshold below which the reduction in covariance determinant results in a warning.

Δt_new :: Time step to be used in the current update.

deterministicforwardmap :: Whether output g comes from a deterministic model.

failed_ens :: Indices of failed particles. If nothing, failures are computed as columns of g with NaN entries.

update_ensemble!(
    ekp::EnsembleKalmanProcess{FT, IT, SparseInversion{FT}},
    g::AbstractMatrix{FT};
    cov_threshold::FT = 0.01,
    Δt_new = nothing,
    deterministic_forward_map = true,
    failed_ens = nothing,
) where {FT, IT}

Updates the ensemble according to a SparseInversion process.

Inputs:

ekp :: The EnsembleKalmanProcess to update.

g :: Model outputs, they need to be stored as a N_obs × N_ens array (i.e data are columms).

cov_threshold :: Threshold below which the reduction in covariance determinant results in a warning.

Δt_new :: Time step to be used in the current update.

deterministicforwardmap :: Whether output g comes from a deterministic model.

failed_ens :: Indices of failed particles. If nothing, failures are computed as columns of g with NaN entries.

update_ensemble!(
    ekp::EnsembleKalmanProcess{FT, IT, Sampler{FT}},
    g::AbstractMatrix{FT};
    failed_ens = nothing,
) where {FT, IT}

Updates the ensemble according to a Sampler process.

Inputs:

ekp :: The EnsembleKalmanProcess to update.

g :: Model outputs, they need to be stored as a N_obs × N_ens array (i.e data are columms).

failed_ens :: Indices of failed particles. If nothing, failures are computed as columns of g with NaN entries.

update_ensemble!(
    uki::EnsembleKalmanProcess{FT, IT, Unscented},
    g_in::AbstractMatrix{FT};
    Δt_new = nothing,
    failed_ens = nothing,
) where {FT <: AbstractFloat, IT <: Int}

Updates the ensemble according to an Unscented process.

Inputs:

uki :: The EnsembleKalmanProcess to update.

gin :: Model outputs, they need to be stored as a `Nobs × N_ens` array (i.e data are columms).

Δt_new :: Time step to be used in the current update.

failed_ens :: Indices of failed particles. If nothing, failures are computed as columns of g with NaN entries.