Localizers

EnsembleKalmanProcesses.Localizers.LocalizationMethod — Type

abstract type LocalizationMethod

Abstract supertype for methods that localize or correct sample covariances in ensemble Kalman updates.

Subtypes:

NoLocalization: no localization or correction.

Delta, RBF, BernoulliDropout: localizing kernels.

SEC, SECFisher, SECNice: sampling-error-correction methods.

EnsembleKalmanProcesses.Localizers.Localizer — Type

Localizer{LM <: LocalizationMethod, T}

Structure that defines a localize function, based on a localization method.

Fields

localize::Function: Localizing function of the form: cov -> kernel .* cov

Constructors

Localizer(localization, J)
Localizer(localization, J, T)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/Localizers.jl:158.

Localizer(localization, J)
Localizer(localization, J, T)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/Localizers.jl:164.

Localizer(localization, J, T)
Localizer(localization, J)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/Localizers.jl:179.

Localizer(localization, J)
Localizer(localization, J, T)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/Localizers.jl:218.

Localizer(localization, J)
Localizer(localization, J, T)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/Localizers.jl:240.

Localizer(localization, J, T)
Localizer(localization, J)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/Localizers.jl:280.

Localizer(localization, J)
Localizer(localization, J, T)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/Localizers.jl:346.

EnsembleKalmanProcesses.Localizers.RBF — Type

RBF{FT <: Real} <: LocalizationMethod

Radial basis function localization method. Covariance terms $C_{i,j}$ are damped through multiplication with a centered Gaussian with standardized deviation $d(i,j)= \vert i-j \vert / l$.

Fields

lengthscale::Real: Length scale defining the RBF kernel

EnsembleKalmanProcesses.Localizers.BernoulliDropout — Type

BernoulliDropout{FT <: Real} <: LocalizationMethod

Localization method that drops cross-covariance terms with probability $1-p$, retaining a Hermitian structure.

Fields

prob::Real: Probability of keeping a given cross-covariance term

EnsembleKalmanProcesses.Localizers.SEC — Type

SEC{FT <: Real} <: LocalizationMethod

Sampling error correction that shrinks correlations by a factor of $\vert r \vert ^\alpha$, as per Lee (2021). Sparsity of the resulting correlations can be imposed through the parameter r_0.

Lee, Y. (2021). Sampling error correction in ensemble Kalman inversion. arXiv:2105.11341 [cs, math]. http://arxiv.org/abs/2105.11341

Fields

α::Real: Controls degree of sampling error correction

r_0::Real: Cutoff correlation

EnsembleKalmanProcesses.Localizers.SECFisher — Type

SECFisher <: LocalizationMethod

Sampling error correction for EKI, as per Lee (2021), but using the method from Flowerdew (2015) based on the Fisher transformation. Correlations are shrunk by a factor determined by the sample correlation and the ensemble size.

Flowerdew, J. (2015). Towards a theory of optimal localisation. Tellus A: Dynamic Meteorology and Oceanography, 67(1), 25257. https://doi.org/10.3402/tellusa.v67.25257

Lee, Y. (2021). Sampling error correction in ensemble Kalman inversion. arXiv:2105.11341 [cs, math]. http://arxiv.org/abs/2105.11341

EnsembleKalmanProcesses.Localizers.SECNice — Type

SECNice{FT <: Real} <: LocalizationMethod

Sampling error correction as of Vishny, Morzfeld, et al. (2024), DOI. Correlations are shrunk by a factor determined by correlation and ensemble size. The factors are automatically determined by a discrepancy principle. Thus no algorithm parameters are required, though some tuning of the discrepancy principle tolerances are made available.

Fields

n_samples::Int64: number of samples to approximate the std of correlation distribution (default 1000)

δ_ug::Real: scaling for discrepancy principle for ug correlation (default 1.0)

δ_gg::Real: scaling for discrepancy principle for gg correlation (default 1.0)

EnsembleKalmanProcesses.Localizers.Delta — Type

Dirac delta localization method, with an identity matrix as the kernel.

EnsembleKalmanProcesses.Localizers.NoLocalization — Type

Idempotent localization method.