ParameterDistribution

EnsembleKalmanProcesses.ParameterDistributionStorage.Parameterized — Type

struct Parameterized <: ParameterDistributionType

A distribution constructed from a parametrized formula (e.g Julia Distributions.jl)

EnsembleKalmanProcesses.ParameterDistributionStorage.Samples — Type

struct Samples{FT<:Real} <: ParameterDistributionType

A distribution comprised of only samples, stored as columns of parameters

EnsembleKalmanProcesses.ParameterDistributionStorage.Constraint — Type

struct Constraint <: ConstraintType

Contains two functions to map between constrained and unconstrained spaces.

EnsembleKalmanProcesses.ParameterDistributionStorage.no_constraint — Function

function no_constraint()

Constructs a Constraint with no constraints, enforced by maps x -> x and x -> x.

EnsembleKalmanProcesses.ParameterDistributionStorage.bounded_below — Function

function bounded_below(lower_bound::FT) where {FT <: Real}

Constructs a Constraint with provided lower bound, enforced by maps x -> log(x - lowerbound) and x -> exp(x) + lowerbound.

EnsembleKalmanProcesses.ParameterDistributionStorage.bounded_above — Function

function bounded_above(upper_bound::FT) where {FT <: Real}

Constructs a Constraint with provided upper bound, enforced by maps x -> log(upperbound - x) and x -> upperbound - exp(x).

EnsembleKalmanProcesses.ParameterDistributionStorage.bounded — Function

function bounded(lower_bound::FT, upper_bound::FT) where {FT <: Real}

Constructs a Constraint with provided upper and lower bounds, enforced by maps x -> log((x - lowerbound) / (upperbound - x)) and x -> (upperbound * exp(x) + lowerbound) / (exp(x) + 1)

EnsembleKalmanProcesses.ParameterDistributionStorage.len — Function

function len(c::Array{CType})

The number of constraints, each constraint has length 1.

EnsembleKalmanProcesses.ParameterDistributionStorage.n_samples — Function

function n_samples(d::Samples)

The number of samples in the array

EnsembleKalmanProcesses.ParameterDistributionStorage.ParameterDistribution — Type

struct ParameterDistribution

Structure to hold a parameter distribution, always stored as an array of distributions

EnsembleKalmanProcesses.ParameterDistributionStorage.get_name — Function

function get_name(pd::ParameterDistribution)

Returns a list of ParameterDistribution names

EnsembleKalmanProcesses.ParameterDistributionStorage.get_dimensions — Function

function get_dimensions(pd::ParameterDistribution)

The number of dimensions of the parameter space

EnsembleKalmanProcesses.ParameterDistributionStorage.get_n_samples — Function

function get_n_samples(pd::ParameterDistribution)

The number of samples in a Samples distribution

EnsembleKalmanProcesses.ParameterDistributionStorage.get_all_constraints — Function

function get_all_constraints(pd::ParameterDistribution)

returns the (flattened) array of constraints of the parameter distribution

EnsembleKalmanProcesses.ParameterDistributionStorage.batch — Function

function batch(pd:ParameterDistribution)

Returns a list of contiguous [collect(1:i), collect(i+1:j),... ] used to split parameter arrays by distribution dimensions

EnsembleKalmanProcesses.ParameterDistributionStorage.get_distribution — Function

function get_distribution(pd::ParameterDistribution)

Returns a Dict of ParameterDistribution distributions, with the parameter names as dictionary keys. For parameters represented by Samples, the samples are returned as a 2D (parameterdimension x nsamples) array

EnsembleKalmanProcesses.ParameterDistributionStorage.sample_distribution — Function

function sample_distribution(pd::ParameterDistribution)

Draws samples from the parameter distributions returns an array, with parameters as columns

EnsembleKalmanProcesses.ParameterDistributionStorage.get_logpdf — Function

function logpdf(pd::ParameterDistribution, xarray::Array{<:Real,1})

Obtains the independent logpdfs of the parameter distributions at xarray (non-Samples Distributions only), and returns their sum.

EnsembleKalmanProcesses.ParameterDistributionStorage.get_cov — Function

function get_cov(pd::ParameterDistribution)

returns a blocked covariance of the distributions

EnsembleKalmanProcesses.ParameterDistributionStorage.get_mean — Function

get_mean(pd::Parameterized)

returns a mean of parameterized distribution

get_mean(pd::Samples)

returns a mean of the samples

function get_mean(pd::ParameterDistribution)

returns a mean of the distributions

EnsembleKalmanProcesses.ParameterDistributionStorage.transform_constrained_to_unconstrained — Function

function transform_constrained_to_unconstrained(pd::ParameterDistribution, x::Array{<:Real})

Apply the transformation to map (possibly constrained) parameters xarray into the unconstrained space

EnsembleKalmanProcesses.ParameterDistributionStorage.transform_unconstrained_to_constrained — Function

function transform_unconstrained_to_constrained(pd::ParameterDistribution, xarray::Array{Real})

Apply the transformation to map parameters xarray from the unconstrained space into (possibly constrained) space