Ensemble Kalman Sampler

EnsembleKalmanProcesses.Sampler — Type

Sampler{FT<:AbstractFloat, T <:SamplerType} <: Process

An ensemble Kalman Sampler process. with type Sampler Type (e.g., ALDI or EKS).

Constructor

Sampler(prior::ParameterDistribution) # ALDI update (samplertype="aldi") Sampler(prior::ParameterDistribution; samplertype = "eks") # EKS update

Fields

prior_mean::Vector{FT} where FT<:AbstractFloat: Mean of Gaussian parameter prior in unconstrained space

prior_cov::Union{LinearAlgebra.UniformScaling{FT}, AbstractMatrix{FT}} where FT<:AbstractFloat: Covariance of Gaussian parameter prior in unconstrained space

Constructors

Sampler(prior; sampler_type)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/EnsembleKalmanSampler.jl:36.

EnsembleKalmanProcesses.eks_update — Function

eks_update(
    ekp::EnsembleKalmanProcess,
    u::AbstractArray{FT<:Real, 2},
    g::AbstractArray{FT<:Real, 2},
    process::Sampler{FT<:Real, EKS}
) -> Any

Returns the updated parameter vectors given their current values and the corresponding forward model evaluations, using the sampler algorithm of (Garbuno-Iñigo Hoffmann Li Stuart 2019)

The current implementation assumes that rows of u and g correspond to ensemble members, so it requires passing the transpose of the u and g arrays associated with ekp.

eks_update(
    ekp::EnsembleKalmanProcess,
    u::AbstractArray{FT<:Real, 2},
    g::AbstractArray{FT<:Real, 2},
    process::Sampler{FT<:Real, ALDI}
) -> Any

Returns the updated parameter vectors given their current values and the corresponding forward model evaluations, using the sampler algorithm of (Garbuno-Iñigo Nüsken Reich 2020)

The current implementation assumes that rows of u and g correspond to ensemble members, so it requires passing the transpose of the u and g arrays associated with ekp.