EnsembleKalmanProcesses

EnsembleKalmanProcesses.EnsembleKalmanProcess — Type

EnsembleKalmanProcess{FT <: AbstractFloat, IT <: Int, P <: Process}

Structure that is used in Ensemble Kalman processes.

Fields

u::Array{EnsembleKalmanProcesses.DataContainers.DataContainer{FT}} where FT<:AbstractFloat

array of stores for parameters (u), each of size [N_par × N_ens]

obs_mean::Vector{FT} where FT<:AbstractFloat

vector of the observed vector size [N_obs]

obs_noise_cov::Union{LinearAlgebra.UniformScaling{FT}, AbstractMatrix{FT}} where FT<:AbstractFloat

covariance matrix of the observational noise, of size [N_obs × N_obs]

N_ens::Int64

ensemble size

g::Array{EnsembleKalmanProcesses.DataContainers.DataContainer{FT}} where FT<:AbstractFloat

Array of stores for forward model outputs, each of size [N_obs × N_ens]

err::Vector{FT} where FT<:AbstractFloat

vector of errors

Δt::Vector{FT} where FT<:AbstractFloat

vector of timesteps used in each EK iteration

process::EnsembleKalmanProcesses.Process

the particular EK process (Inversion or Sampler or Unscented or SparseInversion)

rng::Random.AbstractRNG

Random number generator object (algorithm + seed) used for sampling and noise, for reproducibility. Defaults to Random.GLOBAL_RNG.

failure_handler::FailureHandler

struct storing failsafe update directives, implemented for (Inversion, SparseInversion, Unscented)

localizer::EnsembleKalmanProcesses.Localizers.Localizer

Localization kernel, implemented for (Inversion, SparseInversion, Unscented)

Constructors

EnsembleKalmanProcess(
    params::AbstractMatrix{FT},
    obs_mean,
    obs_noise_cov::Union{AbstractMatrix{FT}, UniformScaling{FT}},
    process::P;
    Δt = FT(1),
    rng::AbstractRNG = Random.GLOBAL_RNG
) where {FT <: AbstractFloat, P <: Process}

Inputs:

params :: Initial parameter ensemble

obs_mean :: Vector of observations

obs_noise_cov :: Noise covariance associated with the observations obs_mean

process :: Algorithm used to evolve the ensemble

Δt :: Initial time step or learning rate

rng :: Random number generator

EnsembleKalmanProcess(u, obs_mean, obs_noise_cov, N_ens, g, err, Δt, process, rng, failure_handler, localizer)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/EnsembleKalmanProcess.jl:88.

EnsembleKalmanProcess(params, obs_mean, obs_noise_cov, process; Δt, rng, failure_handler_method, localization_method)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/EnsembleKalmanProcess.jl:112.

EnsembleKalmanProcess(obs_mean, obs_noise_cov, process; Δt, rng, failure_handler_method, localization_method)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/UnscentedKalmanInversion.jl:132.

EnsembleKalmanProcesses.Inversion — Type

Inversion <: Process

An ensemble Kalman Inversion process

EnsembleKalmanProcesses.SparseInversion — Type

SparseInversion <: Process

A sparse ensemble Kalman Inversion process

Fields

γ::AbstractFloat

upper limit of l1-norm

threshold_value::AbstractFloat

threshold below which the norm of parameters is pruned to zero

uc_idx::Union{Colon, AbstractVector}

indices of parameters included in the evaluation of l1-norm constraint

reg::AbstractFloat

a small regularization value to enhance robustness of convex optimization

Constructors

SparseInversion(γ, threshold_value, uc_idx, reg)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/SparseEnsembleKalmanInversion.jl:19.

SparseInversion(γ; threshold_value, uc_idx, reg)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/SparseEnsembleKalmanInversion.jl:29.

SparseInversion(; γ, threshold_value, uc_idx, reg)

EnsembleKalmanProcesses.Unscented — Type

Unscented{FT<:AbstractFloat, IT<:Int} <: Process

An unscented Kalman Inversion process.

Fields

u_mean::Any

an interable of arrays of size N_parameters containing the mean of the parameters (in each uki iteration a new array of mean is added)

uu_cov::Any

an iterable of arrays of size (N_parameters x N_parameters) containing the covariance of the parameters (in each uki iteration a new array of cov is added)

obs_pred::Any

an iterable of arrays of size N_y containing the predicted observation (in each uki iteration a new array of predicted observation is added)

c_weights::AbstractVector{FT} where FT<:AbstractFloat

weights in UKI

mean_weights::AbstractVector{FT} where FT<:AbstractFloat

cov_weights::AbstractVector{FT} where FT<:AbstractFloat

Σ_ω::AbstractMatrix{FT} where FT<:AbstractFloat

covariance of the artificial evolution error

Σ_ν_scale::AbstractFloat

covariance of the artificial observation error

α_reg::AbstractFloat

regularization parameter

r::AbstractVector{FT} where FT<:AbstractFloat

regularization vector

update_freq::Int64

update frequency

iter::Int64

current iteration number

Constructors

Unscented(
    u0_mean::AbstractVector{FT},
    uu0_cov::AbstractMatrix{FT};
    α_reg::FT = 1.0,
    update_freq::IT = 1,
    modified_unscented_transform::Bool = true,
    prior_mean::Union{AbstractVector{FT}, Nothing} = nothing,
) where {FT <: AbstractFloat, IT <: Int}

Construct an Unscented Inversion Process.

Inputs:

u0_mean: Mean at initialization.

uu0_cov: Covariance at initialization.

α_reg: Hyperparameter controlling regularization toward the prior mean (0 < α_reg ≤ 1),

default should be 1, without regulariazion.

update_freq: Set to 0 when the inverse problem is not identifiable,

namely the inverse problem has multiple solutions, the covariance matrix will represent only the sensitivity of the parameters, instead of posterior covariance information; set to 1 (or anything > 0) when the inverse problem is identifiable, and the covariance matrix will converge to a good approximation of the posterior covariance with an uninformative prior.

modified_unscented_transform: Modification of the UKI quadrature given in Huang et al (2021).

prior_mean: Prior mean used for regularization.

Unscented(u_mean, uu_cov, obs_pred, c_weights, mean_weights, cov_weights, Σ_ω, Σ_ν_scale, α_reg, r, update_freq, iter)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/UnscentedKalmanInversion.jl:45.

Unscented(u0_mean, uu0_cov; α_reg, update_freq, modified_unscented_transform, prior_mean)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/UnscentedKalmanInversion.jl:69.

EnsembleKalmanProcesses.Sampler — Type

Sampler{FT<:AbstractFloat,IT<:Int} <: Process

An ensemble Kalman Sampler process.

Fields

prior_mean::Vector{FT} where FT<:AbstractFloat

Mean of Gaussian parameter prior in unconstrained space

prior_cov::AbstractMatrix{FT} where FT<:AbstractFloat

Covariance of Gaussian parameter prior in unconstrained space

Constructors

Sampler(prior_mean, prior_cov)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/EnsembleKalmanSampler.jl:17.

EnsembleKalmanProcesses.FailureHandler — Type

FailureHandler{P <: Process, FM <: FailureHandlingMethod}

Structure defining the failure handler method used in the EnsembleKalmanProcess.

Fields

failsafe_update::Function

Failsafe algorithmic update equation

Constructors

FailureHandler(process, method)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/EnsembleKalmanInversion.jl:10.

FailureHandler(process, method)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/EnsembleKalmanInversion.jl:22.

FailureHandler(process, method)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/EnsembleKalmanSampler.jl:23.

FailureHandler(process, method)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/SparseEnsembleKalmanInversion.jl:38.

FailureHandler(process, method)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/SparseEnsembleKalmanInversion.jl:50.

FailureHandler(process, method)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/UnscentedKalmanInversion.jl:175.

FailureHandler(process, method)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/UnscentedKalmanInversion.jl:194.

EnsembleKalmanProcesses.SampleSuccGauss — Type

" SampleSuccGauss <: FailureHandlingMethod

Failure handling method that substitutes failed ensemble members by new samples from the empirical Gaussian distribution defined by the updated successful ensemble.

EnsembleKalmanProcesses.IgnoreFailures — Type

Failure handling method that ignores forward model failures

EnsembleKalmanProcesses.get_u — Function

get_u(ekp::EnsembleKalmanProcess, iteration::IT; return_array=true) where {IT <: Integer}

Get for the EKI iteration. Returns a DataContainer object unless array is specified.

get_u(ekp::EnsembleKalmanProcess; return_array=true)

Get for the EKI iteration. Returns a DataContainer object unless array is specified.

EnsembleKalmanProcesses.get_g — Function

get_g(ekp::EnsembleKalmanProcess, iteration::IT; return_array=true) where {IT <: Integer}

Get for the EKI iteration. Returns a DataContainer object unless array is specified.

get_g(ekp::EnsembleKalmanProcess; return_array=true)

Get for the EKI iteration. Returns a DataContainer object unless array is specified.

EnsembleKalmanProcesses.get_u_final — Function

get_u_final(ekp::EnsembleKalmanProcess, return_array=true)

Get the final or prior iteration of parameters or model ouputs, returns a DataContainer Object if return_array is false.

EnsembleKalmanProcesses.compute_error! — Function

compute_error!(ekp::EnsembleKalmanProcess)

Computes the covariance-weighted error of the mean forward model output, (ḡ - y)'Γ_inv(ḡ - y). The error is stored within the EnsembleKalmanProcess.

EnsembleKalmanProcesses.get_error — Function

get_error(ekp::EnsembleKalmanProcess)

Returns the mean forward model output error as a function of algorithmic time.

EnsembleKalmanProcesses.get_N_iterations — Function

get_N_iterations(ekp::EnsembleKalmanProcess

Get number of times update has been called (equals size(g), or size(u)-1).

EnsembleKalmanProcesses.construct_initial_ensemble — Function

construct_initial_ensemble(
    rng::AbstractRNG,
    prior::ParameterDistribution,
    N_ens::IT;
    rng_seed::Union{IT, Nothing} = nothing,
) where {IT <: Int}
construct_initial_ensemble(prior::ParameterDistribution, N_ens::IT; kwargs...) where {IT <: Int}

Construct the initial parameters, by sampling N_ens samples from specified prior distribution. Returned with parameters as columns.

EnsembleKalmanProcesses.find_ekp_stepsize — Function

find_ekp_stepsize(
    ekp::EnsembleKalmanProcess{FT, IT, Inversion},
    g::AbstractMatrix{FT};
    cov_threshold::FT = 0.01,
) where {FT, IT}

Find largest stepsize for the EK solver that leads to a reduction of the determinant of the sample covariance matrix no greater than cov_threshold.

EnsembleKalmanProcesses.update_ensemble! — Function

update_ensemble!(
    ekp::EnsembleKalmanProcess{FT, IT, Inversion},
    g::AbstractMatrix{FT};
    cov_threshold::FT = 0.01,
    Δt_new::Union{Nothing, FT} = nothing,
    deterministic_forward_map::Bool = true,
    failed_ens = nothing,
) where {FT, IT}

Updates the ensemble according to an Inversion process.

Inputs:

ekp :: The EnsembleKalmanProcess to update.

g :: Model outputs, they need to be stored as a N_obs × N_ens array (i.e data are columms).

cov_threshold :: Threshold below which the reduction in covariance determinant results in a warning.

Δt_new :: Time step to be used in the current update.

deterministicforwardmap :: Whether output g comes from a deterministic model.

failed_ens :: Indices of failed particles. If nothing, failures are computed as columns of g with NaN entries.

update_ensemble!(
    ekp::EnsembleKalmanProcess{FT, IT, SparseInversion{FT}},
    g::AbstractMatrix{FT};
    cov_threshold::FT = 0.01,
    Δt_new = nothing,
    deterministic_forward_map = true,
    failed_ens = nothing,
) where {FT, IT}

Updates the ensemble according to a SparseInversion process.

Inputs:

ekp :: The EnsembleKalmanProcess to update.

g :: Model outputs, they need to be stored as a N_obs × N_ens array (i.e data are columms).

cov_threshold :: Threshold below which the reduction in covariance determinant results in a warning.

Δt_new :: Time step to be used in the current update.

deterministicforwardmap :: Whether output g comes from a deterministic model.

failed_ens :: Indices of failed particles. If nothing, failures are computed as columns of g with NaN entries.

update_ensemble!(
    ekp::EnsembleKalmanProcess{FT, IT, Sampler{FT}},
    g::AbstractMatrix{FT};
    failed_ens = nothing,
) where {FT, IT}

Updates the ensemble according to a Sampler process.

Inputs:

ekp :: The EnsembleKalmanProcess to update.

g :: Model outputs, they need to be stored as a N_obs × N_ens array (i.e data are columms).

failed_ens :: Indices of failed particles. If nothing, failures are computed as columns of g with NaN entries.

update_ensemble!(
    uki::EnsembleKalmanProcess{FT, IT, Unscented},
    g_in::AbstractMatrix{FT};
    Δt_new = nothing,
    failed_ens = nothing,
) where {FT <: AbstractFloat, IT <: Int}

Updates the ensemble according to an Unscented process.

Inputs:

uki :: The EnsembleKalmanProcess to update.

gin :: Model outputs, they need to be stored as a `Nobs × N_ens` array (i.e data are columms).

Δt_new :: Time step to be used in the current update.

failed_ens :: Indices of failed particles. If nothing, failures are computed as columns of g with NaN entries.

EnsembleKalmanProcesses.sample_empirical_gaussian — Function

sample_empirical_gaussian(
    u::AbstractMatrix{FT},
    n::IT;
    inflation::Union{FT, Nothing} = nothing,
) where {FT <: Real, IT <: Int}

Returns n samples from an empirical Gaussian based on point estimates u, adding inflation if the covariance is singular.

EnsembleKalmanProcesses.split_indices_by_success — Function

 split_indices_by_success(g::AbstractMatrix{FT}) where {FT <: Real}

Returns the successful/failed particle indices given a matrix with output vectors stored as columns. Failures are defined for particles containing at least one NaN output element.

EnsembleKalmanProcesses.eki_update — Function

 eki_update(
    ekp::EnsembleKalmanProcess{FT, IT, Inversion},
    u::AbstractMatrix{FT},
    g::AbstractMatrix{FT},
    y::AbstractMatrix{FT},
    obs_noise_cov::Union{AbstractMatrix{FT}, UniformScaling{FT}},
) where {FT <: Real, IT}

Returns the updated parameter vectors given their current values and the corresponding forward model evaluations, using the inversion algorithm from eqns. (4) and (5) of Schillings and Stuart (2017).

Localization is applied following Tong and Morzfeld (2022).

EnsembleKalmanProcesses.sparse_eki_update — Function

 sparse_eki_update(
    ekp::EnsembleKalmanProcess{FT, IT, SparseInversion{FT}},
    u::AbstractMatrix{FT},
    g::AbstractMatrix{FT},
    y::AbstractMatrix{FT},
    obs_noise_cov::Union{AbstractMatrix{FT}, UniformScaling{FT}},
) where {FT <: Real, IT}

Returns the sparse updated parameter vectors given their current values and the corresponding forward model evaluations, using the inversion algorithm from eqns. (3.7) and (3.14) of Schneider et al. (2021).

Localization is applied following Tong and Morzfeld (2022).

EnsembleKalmanProcesses.eks_update — Function

 eks_update(
    ekp::EnsembleKalmanProcess{FT, IT, Sampler{FT}},
    u::AbstractMatrix{FT},
    g::AbstractMatrix{FT},
) where {FT <: Real, IT}

Returns the updated parameter vectors given their current values and the corresponding forward model evaluations, using the sampler algorithm.

The current implementation assumes that rows of u and g correspond to ensemble members, so it requires passing the transpose of the u and g arrays associated with ekp.

EnsembleKalmanProcesses.sparse_qp — Function

sparse_qp(
    ekp::EnsembleKalmanProcess{FT, IT, SparseInversion{FT}},
    v_j::Vector{FT},
    cov_vv_inv::AbstractMatrix{FT},
    H_u::SparseArrays.SparseMatrixCSC{FT},
    H_g::SparseArrays.SparseMatrixCSC{FT},
    y_j::Vector{FT};
    H_uc::SparseArrays.SparseMatrixCSC{FT} = H_u,
) where {FT, IT}

Solving quadratic programming problem with sparsity constraint.