Observations

EnsembleKalmanProcesses.Observations.Obs — Type

Obs{FT<:AbstractFloat}

Structure that contains the observations

Fields

samples

vector of observational samples, each of length sample_dim

obs_noise_cov

covariance of the observational noise (assumed to be normally distributed); sampledim x sampledim (where sampledim is the number of elements in each sample), or a scalar if the sample dim is 1. If not supplied, obsnoisecov is set to a diagonal matrix whose non-zero elements are the variances of the samples, or to a scalar variance in the case of 1d samples. obsnoise_cov is set to nothing if only a single sample is provided.

mean

sample mean

data_names

names of the data