Observations

EnsembleKalmanProcesses.Observations.Observation — Type

Observation{FT <: AbstractFloat}

Structure that contains the observations

Fields

samples::Array{Vector{FT}, 1} where FT<:AbstractFloat

vector of observational samples, each of length sample_dim

obs_noise_cov::Union{Nothing, LinearAlgebra.UniformScaling{FT}, AbstractMatrix{FT}, FT} where FT<:AbstractFloat

covariance of the observational noise (assumed to be normally distributed); sample_dim x sample_dim (where sample_dim is the number of elements in each sample), or a scalar if the sample dim is 1. If not supplied, obs_noise_cov is set to a diagonal matrix whose non-zero elements are the variances of the samples, or to a scalar variance in the case of 1d samples. obs_noise_cov is set to nothing if only a single sample is provided.

mean::Union{AbstractVector{FT}, FT} where FT<:AbstractFloat

sample mean

data_names::Union{String, AbstractVector{String}}

names of the data