ParameterDistributions

EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution — Type

ParameterDistribution

Structure to hold a parameter distribution, always stored as an array of distributions internally.

Fields

distribution::AbstractVector{PDType} where PDType<:EnsembleKalmanProcesses.ParameterDistributions.ParameterDistributionType

Vector of parameter distributions, defined in unconstrained space

constraint::AbstractVector{CType} where CType<:EnsembleKalmanProcesses.ParameterDistributions.ConstraintType

Vector of constraints defining transformations between constrained and unconstrained space

name::AbstractVector{ST} where ST<:AbstractString

Vector of parameter names

Constructors

ParameterDistribution(distribution, constraint, name)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/ParameterDistributions.jl:270.

ParameterDistribution(param_dist_dict)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/ParameterDistributions.jl:286.

ParameterDistribution(distribution, constraint, name)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/ParameterDistributions.jl:368.

EnsembleKalmanProcesses.ParameterDistributions.constrained_gaussian — Function

constrained_gaussian(
    name::AbstractString,
    μ_c::Real,
    σ_c::Real,
    lower_bound::Real,
    upper_bound::Real;
    repeats = 1,
    optim_algorithm::Optim.AbstractOptimizer = NelderMead(),
    optim_kwargs...,
)

Constructor for a 1D ParameterDistribution consisting of a transformed Gaussian, constrained to have support on [lower_bound, upper_bound], with first two moments μ_c and σ_c^2. The moment integrals can't be done in closed form, so we set the parameters of the distribution with numerical optimization.

Note

The intended use case is defining priors set from user expertise for use in inference with adequate data, so for the sake of performance we only require that the optimization reproduce μ_c, σ_c to a loose tolerance (1e-5). Warnings are logged when the optimization fails.

Note

The distribution may be bimodal for σ_c large relative to the width of the bound interval. In extreme cases the distribution becomes concentrated at the bound endpoints. We regard this as a feature, not a bug, and do not warn the user when bimodality occurs.

EnsembleKalmanProcesses.ParameterDistributions.Parameterized — Type

Parameterized <: ParameterDistributionType

A distribution constructed from a parameterized formula (e.g Julia Distributions.jl)

Fields

distribution::Distributions.Distribution

A parameterized distribution

EnsembleKalmanProcesses.ParameterDistributions.Samples — Type

Samples{FT <: Real} <: ParameterDistributionType

A distribution comprised of only samples, stored as columns of parameters.

Fields

distribution_samples::AbstractMatrix{FT} where FT<:Real

Samples defining an empirical distribution, stored as columns

EnsembleKalmanProcesses.ParameterDistributions.Constraint — Type

Constraint{T} <: ConstraintType

Class describing a 1D bijection between constrained and unconstrained spaces. Included parametric types for T:

NoConstraint

BoundedBelow

BoundedAbove

Bounded

Fields

constrained_to_unconstrained::Function

A map from constrained domain -> (-Inf,Inf)

c_to_u_jacobian::Function

The jacobian of the map from constrained domain -> (-Inf,Inf)

unconstrained_to_constrained::Function

Map from (-Inf,Inf) -> constrained domain

bounds::Union{Nothing, Dict}

Dictionary of values used to build the Constraint (e.g. "lowerbound" or "upperbound")

EnsembleKalmanProcesses.ParameterDistributions.no_constraint — Function

no_constraint()

Constructs a Constraint with no constraints, enforced by maps x -> x and x -> x.

EnsembleKalmanProcesses.ParameterDistributions.bounded_below — Function

bounded_below(lower_bound::FT) where {FT <: Real}

Constructs a Constraint with provided lower bound, enforced by maps x -> log(x - lower_bound) and x -> exp(x) + lower_bound.

EnsembleKalmanProcesses.ParameterDistributions.bounded_above — Function

bounded_above(upper_bound::FT) where {FT <: Real}

Constructs a Constraint with provided upper bound, enforced by maps x -> log(upper_bound - x) and x -> upper_bound - exp(x).

EnsembleKalmanProcesses.ParameterDistributions.bounded — Function

bounded(lower_bound::Real, upper_bound::Real)

Constructs a Constraint with provided upper and lower bounds, enforced by maps x -> log((x - lower_bound) / (upper_bound - x)) and x -> (upper_bound * exp(x) + lower_bound) / (exp(x) + 1).

Base.length — Method

length(c<:ConstraintType)

A constraint has length 1.

Base.size — Method

size(c<:ConstraintType)

A constraint has size 1.

EnsembleKalmanProcesses.ParameterDistributions.n_samples — Function

n_samples(d<:Samples)

The number of samples in the array.

EnsembleKalmanProcesses.ParameterDistributions.get_name — Function

get_name(pd::ParameterDistribution)

Returns a list of ParameterDistribution names.

EnsembleKalmanProcesses.ParameterDistributions.get_dimensions — Function

get_dimensions(pd::ParameterDistribution)

The number of dimensions of the parameter space.

EnsembleKalmanProcesses.ParameterDistributions.get_n_samples — Function

get_n_samples(pd::ParameterDistribution)

The number of samples in a Samples distribution

EnsembleKalmanProcesses.ParameterDistributions.get_all_constraints — Function

get_all_constraints(pd::ParameterDistribution)

Returns the (flattened) array of constraints of the parameter distribution.

EnsembleKalmanProcesses.ParameterDistributions.get_bounds — Function

get_bounds(c::Constraint)

Gets the bounds field from the Constraint.

EnsembleKalmanProcesses.ParameterDistributions.get_constraint_type — Function

get_bounds(c::Constraint{T})

Gets the parametric type T.

EnsembleKalmanProcesses.ParameterDistributions.batch — Function

batch(pd::ParameterDistribution)

Returns a list of contiguous [collect(1:i), collect(i+1:j),... ] used to split parameter arrays by distribution dimensions.

EnsembleKalmanProcesses.ParameterDistributions.get_distribution — Function

get_distribution(pd::ParameterDistribution)

Returns a Dict of ParameterDistribution distributions, with the parameter names as dictionary keys. For parameters represented by Samples, the samples are returned as a 2D (parameter_dimension x n_samples) array.

StatsBase.sample — Function

sample([rng], pd::ParameterDistribution, [n_draws])

Draws n_draws samples from the parameter distributions pd. Returns an array, with parameters as columns. rng is optional and defaults to Random.GLOBAL_RNG. n_draws is optional and defaults to 1.

sample([rng], d::Samples, [n_draws])

Draws n_draws samples from the parameter distributions d. Returns an array, with parameters as columns. rng is optional and defaults to Random.GLOBAL_RNG. n_draws is optional and defaults to 1.

sample([rng], d::Parameterized, [n_draws])

Draws n_draws samples from the parameter distributions d. Returns an array, with parameters as columns. rng is optional and defaults to Random.GLOBAL_RNG. n_draws is optional and defaults to 1.

sample([rng], d::VectorOfParameterized, [n_draws])

Draws n_draws samples from the parameter distributions d. Returns an array, with parameters as columns. rng is optional and defaults to Random.GLOBAL_RNG. n_draws is optional and defaults to 1.

EnsembleKalmanProcesses.ParameterDistributions.get_logpdf — Function

logpdf(pd::ParameterDistribution, xarray::Array{<:Real,1})

Obtains the independent logpdfs of the parameter distributions at xarray (non-Samples Distributions only), and returns their sum.

Statistics.mean — Function

mean(pd::ParameterDistribution)

Returns a concatenated mean of the parameter distributions.

Statistics.var — Function

var(pd::ParameterDistribution)

Returns a flattened variance of the distributions

Statistics.cov — Function

cov(pd::ParameterDistribution)

Returns a dense blocked (co)variance of the distributions.

EnsembleKalmanProcesses.ParameterDistributions.transform_constrained_to_unconstrained — Function

transform_constrained_to_unconstrained(pd::ParameterDistribution, xarray::Array{<:Real,1})

Apply the transformation to map (possibly constrained) parameters xarray into the unconstrained space.

transform_constrained_to_unconstrained(pd::ParameterDistribution, xarray::Array{<:Real,2})

Apply the transformation to map (possibly constrained) parameter samples xarray into the unconstrained space. Here, xarray contains parameters as columns and samples as rows.

EnsembleKalmanProcesses.ParameterDistributions.transform_unconstrained_to_constrained — Function

transform_unconstrained_to_constrained(pd::ParameterDistribution, xarray::Array{<:Real,1})

Apply the transformation to map parameters xarray from the unconstrained space into (possibly constrained) space.

transform_unconstrained_to_constrained(pd::ParameterDistribution, xarray::Array{<:Real,2})

Apply the transformation to map parameter samples xarray from the unconstrained space into (possibly constrained) space. Here, xarray contains parameters as columns and samples as rows.

transform_unconstrained_to_constrained(pd::ParameterDistribution, xarray::Array{Array{<:Real,2},1})

Apply the transformation to map parameter sample ensembles xarray from the unconstrained space into (possibly constrained) space. Here, xarray is an iterable of parameters sample ensembles for different EKP iterations.