ParameterDistributions

EnsembleKalmanProcesses.ParameterDistributions.Parameterized — Type

Parameterized <: ParameterDistributionType

A distribution constructed from a parametrized formula (e.g Julia Distributions.jl)

EnsembleKalmanProcesses.ParameterDistributions.Samples — Type

Samples{FT <: Real} <: ParameterDistributionType

A distribution comprised of only samples, stored as columns of parameters.

EnsembleKalmanProcesses.ParameterDistributions.Constraint — Type

Constraint <: ConstraintType

Contains two functions to map between constrained and unconstrained spaces.

EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution — Type

ParameterDistribution

Structure to hold a parameter distribution, always stored as an array of distributions.

EnsembleKalmanProcesses.ParameterDistributions.no_constraint — Function

no_constraint()

Constructs a Constraint with no constraints, enforced by maps x -> x and x -> x.

EnsembleKalmanProcesses.ParameterDistributions.bounded_below — Function

bounded_below(lower_bound::FT) where {FT <: Real}

Constructs a Constraint with provided lower bound, enforced by maps x -> log(x - lower_bound) and x -> exp(x) + lower_bound.

EnsembleKalmanProcesses.ParameterDistributions.bounded_above — Function

bounded_above(upper_bound::FT) where {FT <: Real}

Constructs a Constraint with provided upper bound, enforced by maps x -> log(upper_bound - x) and x -> upper_bound - exp(x).

EnsembleKalmanProcesses.ParameterDistributions.bounded — Function

bounded(lower_bound::FT, upper_bound::FT) where {FT <: Real}

Constructs a Constraint with provided upper and lower bounds, enforced by maps x -> log((x - lower_bound) / (upper_bound - x)) and x -> (upper_bound * exp(x) + lower_bound) / (exp(x) + 1).

Base.length — Method

length(c<:ConstraintType)

A constraint has length 1.

Base.size — Method

size(c<:ConstraintType)

A constraint has size 1.

EnsembleKalmanProcesses.ParameterDistributions.n_samples — Function

n_samples(d<:Samples)

The number of samples in the array.

EnsembleKalmanProcesses.ParameterDistributions.get_name — Function

get_name(pd::ParameterDistribution)

Returns a list of ParameterDistribution names.

EnsembleKalmanProcesses.ParameterDistributions.get_dimensions — Function

get_dimensions(pd::ParameterDistribution)

The number of dimensions of the parameter space.

EnsembleKalmanProcesses.ParameterDistributions.get_n_samples — Function

get_n_samples(pd::ParameterDistribution)

The number of samples in a Samples distribution

EnsembleKalmanProcesses.ParameterDistributions.get_all_constraints — Function

get_all_constraints(pd::ParameterDistribution)

Returns the (flattened) array of constraints of the parameter distribution.

EnsembleKalmanProcesses.ParameterDistributions.batch — Function

batch(pd::ParameterDistribution)

Returns a list of contiguous [collect(1:i), collect(i+1:j),... ] used to split parameter arrays by distribution dimensions.

EnsembleKalmanProcesses.ParameterDistributions.get_distribution — Function

get_distribution(pd::ParameterDistribution)

Returns a Dict of ParameterDistribution distributions, with the parameter names as dictionary keys. For parameters represented by Samples, the samples are returned as a 2D (parameter_dimension x n_samples) array.

StatsBase.sample — Function

sample([rng], pd::ParameterDistribution, [n_draws])

Draws n_draws samples from the parameter distributions pd. Returns an array, with parameters as columns. rng is optional and defaults to Random.GLOBAL_RNG. n_draws is optional and defaults to 1.

sample([rng], d::Samples, [n_draws])

Draws n_draws samples from the parameter distributions d. Returns an array, with parameters as columns. rng is optional and defaults to Random.GLOBAL_RNG. n_draws is optional and defaults to 1.

sample([rng], d::Parameterized, [n_draws])

Draws n_draws samples from the parameter distributions d. Returns an array, with parameters as columns. rng is optional and defaults to Random.GLOBAL_RNG. n_draws is optional and defaults to 1.

EnsembleKalmanProcesses.ParameterDistributions.get_logpdf — Function

logpdf(pd::ParameterDistribution, xarray::Array{<:Real,1})

Obtains the independent logpdfs of the parameter distributions at xarray (non-Samples Distributions only), and returns their sum.

Statistics.mean — Function

mean(pd::ParameterDistribution)

Returns a concatenated mean of the parameter distributions.

Statistics.var — Function

var(pd::ParameterDistribution)

Returns a flattened variance of the distributions

Statistics.cov — Function

cov(pd::ParameterDistribution)

Returns a dense blocked (co)variance of the distributions.

EnsembleKalmanProcesses.ParameterDistributions.transform_constrained_to_unconstrained — Function

transform_constrained_to_unconstrained(pd::ParameterDistribution, xarray::Array{<:Real,1})

Apply the transformation to map (possibly constrained) parameters xarray into the unconstrained space.

transform_constrained_to_unconstrained(pd::ParameterDistribution, xarray::Array{<:Real,2})

Apply the transformation to map (possibly constrained) parameter samples xarray into the unconstrained space. Here, xarray contains parameters as columns and samples as rows.

EnsembleKalmanProcesses.ParameterDistributions.transform_unconstrained_to_constrained — Function

transform_unconstrained_to_constrained(pd::ParameterDistribution, xarray::Array{<:Real,1})

Apply the transformation to map parameters xarray from the unconstrained space into (possibly constrained) space.

transform_unconstrained_to_constrained(pd::ParameterDistribution, xarray::Array{<:Real,2})

Apply the transformation to map parameter samples xarray from the unconstrained space into (possibly constrained) space. Here, xarray contains parameters as columns and samples as rows.

transform_unconstrained_to_constrained(pd::ParameterDistribution, xarray::Array{Array{<:Real,2},1})

Apply the transformation to map parameter sample ensembles xarray from the unconstrained space into (possibly constrained) space. Here, xarray is an iterable of parameters sample ensembles for different EKP iterations.