Ensemble Kalman Sampler

EnsembleKalmanProcesses.Sampler — Type

Sampler{FT<:AbstractFloat,IT<:Int} <: Process

An ensemble Kalman Sampler process.

Fields

prior_mean::Vector{FT} where FT<:AbstractFloat

Mean of Gaussian parameter prior in unconstrained space

prior_cov::Union{LinearAlgebra.UniformScaling{FT}, AbstractMatrix{FT}} where FT<:AbstractFloat

Covariance of Gaussian parameter prior in unconstrained space

Constructors

Sampler(prior_mean, prior_cov)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/EnsembleKalmanSampler.jl:17.

Sampler(prior)

defined at /home/runner/work/EnsembleKalmanProcesses.jl/EnsembleKalmanProcesses.jl/src/EnsembleKalmanSampler.jl:23.

EnsembleKalmanProcesses.eks_update — Function

 eks_update(
    ekp::EnsembleKalmanProcess{FT, IT, Sampler{FT}},
    u::AbstractMatrix{FT},
    g::AbstractMatrix{FT},
) where {FT <: Real, IT}

Returns the updated parameter vectors given their current values and the corresponding forward model evaluations, using the sampler algorithm.

The current implementation assumes that rows of u and g correspond to ensemble members, so it requires passing the transpose of the u and g arrays associated with ekp.