Visualize

Add a Makie-backend package to your Project.toml

Import one of the Makie backends (GLMakie, CairoMakie, WGLMakie, RPRMakie, etc.) to enable these functions!

EnsembleKalmanProcesses.Visualize.plot_parameter_distribution — Function

Visualize.plot_parameter_distribution(fig::Union{Makie.Figure, Makie.GridLayout, Makie.GridPosition, Makie.GridSubposition},
                       pd::ParameterDistribution;
                       constrained = true,
                       n_sample = 1e4,
                       rng = Random.GLOBAL_RNG)

Plot the distributions pd on fig.

Visualize.plot_parameter_distribution(fig::Union{Makie.Figure, Makie.GridLayout},
                       pd::PDT;
                       constrained = true,
                       n_sample = 1e4,
                       rng = Random.GLOBAL_RNG)
                       where {PDT <: ParameterDistributionType}

Plot the distribution on fig.

EnsembleKalmanProcesses.Visualize.plot_error_over_iters — Function

Visualize.plot_error_over_iters(gridposition, ekp; kwargs...)

Plot the errors from ekp against the number of iterations on gridposition.

Any keyword arguments is passed to the plotting function which takes in any keyword arguments supported by Makie.Lines.

EnsembleKalmanProcesses.Visualize.plot_error_over_iters! — Function

Visualize.plot_error_over_iters!(axis, ekp; kwargs...)

Plot the errors from ekp against the number of iterations on axis.

Any keyword arguments is passed to the plotting function which takes in any keyword arguments supported by Makie.Lines.

EnsembleKalmanProcesses.Visualize.plot_error_over_time — Function

Visualize.plot_error_over_time(gridposition, ekp; kwargs...)

Plot the errors from ekp against time on gridposition.

Any keyword arguments is passed to the plotting function which takes in any keyword arguments supported by Makie.Lines.

EnsembleKalmanProcesses.Visualize.plot_error_over_time! — Function

Visualize.plot_error_over_time!(axis, ekp; kwargs...)

Plot the errors from ekp against time on axis.

Any keyword arguments is passed to the plotting function which takes in any keyword arguments supported by Makie.Lines.

EnsembleKalmanProcesses.Visualize.plot_ϕ_over_iters — Function

Visualize.plot_ϕ_over_iters(gridposition, ekp, prior, dim_idx; kwargs...)

Plot the constrained parameter of index dim_idx against time on gridposition.

Any keyword arguments is passed to the plotting function which takes in any keyword arguments supported by Makie.Scatter.

Visualize.plot_ϕ_over_iters(gridpositions, ekp, prior, name; kwargs...)

Plot the constrained parameter belonging to distribution name against the number of iterations on the iterable gridpositions.

Any keyword arguments are passed to all plots.

EnsembleKalmanProcesses.Visualize.plot_ϕ_over_iters! — Function

Visualize.plot_ϕ_over_iters!(axis, ekp, prior, dim_idx; kwargs...)

Plot the constrained parameter of index dim_idx against time on axis.

Any keyword arguments is passed to the plotting function which takes in any keyword arguments supported by Makie.Scatter.

EnsembleKalmanProcesses.Visualize.plot_ϕ_over_time — Function

Visualize.plot_ϕ_over_time(gridposition, ekp, prior, dim_idx; kwargs...)

Plot the constrained parameter of index dim_idx against time on gridposition.

Any keyword arguments is passed to the plotting function which takes in any keyword arguments supported by Makie.Scatter.

Visualize.plot_ϕ_over_time(gridpositions, ekp, prior, name; kwargs...)

Plot the constrained parameter belonging to distribution name against time on the iterable gridpositions.

Any keyword arguments are passed to all plots.

EnsembleKalmanProcesses.Visualize.plot_ϕ_over_time! — Function

Visualize.plot_ϕ_over_time!(axis, ekp, prior, dim_idx; kwargs...)

Plot the constrained parameter of index dim_idx against time on axis.

Any keyword arguments is passed to the plotting function which takes in any keyword arguments supported by Makie.Scatter.

EnsembleKalmanProcesses.Visualize.plot_ϕ_mean_over_iters! — Function

Visualize.plot_ϕ_mean_over_iters!(axis, ekp, prior, dim_idx; plot_std = false, kwargs...)

Plot the mean constrained parameter of index dim_idx of prior against the number of iterations on axis.

If plot_std = true, then the standard deviation of the constrained parameters of the ensemble is also plotted.

Any keyword arguments is passed to the plotting function which takes in any keyword arguments supported by Makie.Lines and Makie.Band if plot_std = true.

Keyword arguments passed to line_kwargs and band_kwargs are merged with kwargs when possible. The keyword arguments in line_kwargs and band_kwargs take priority over the keyword arguments in kwargs.

EnsembleKalmanProcesses.Visualize.plot_ϕ_mean_over_iters — Function

Visualize.plot_ϕ_mean_over_iters(gridposition, ekp, prior, dim_idx; plot_std = false, kwargs...)

Plot the mean constrained parameter of index dim_idx of prior against the number of iterations on gridposition.

If plot_std = true, then the standard deviation of the constrained parameters of the ensemble is also plotted.

Any keyword arguments is passed to the plotting function which takes in any keyword arguments supported by Makie.Lines and Makie.Band if plot_std = true.

Keyword arguments passed to line_kwargs and band_kwargs are merged with kwargs when possible. The keyword arguments in line_kwargs and band_kwargs take priority over the keyword arguments in kwargs.

Visualize.plot_ϕ_mean_over_iters(gridpositions, ekp::EnsembleKalmanProcess, prior, name; kwargs...)

Plot the mean constrained parameter belonging to distribution name against the number of iterations on the iterable gridpositions.

Any keyword arguments are passed to all plots.

EnsembleKalmanProcesses.Visualize.plot_ϕ_mean_over_time! — Function

Visualize.plot_ϕ_mean_over_time!(axis, ekp, prior, dim_idx; plot_std = false, kwargs...)

Plot the mean constrained parameter of index dim_idx of prior against time on axis.

If plot_std = true, then the standard deviation of the constrained parameter of the ensemble is also plotted.

Any keyword arguments is passed to the plotting function which takes in any keyword arguments supported by Makie.Lines and Makie.Band if plot_std = true.

Keyword arguments passed to line_kwargs and band_kwargs are merged with kwargs when possible. The keyword arguments in line_kwargs and band_kwargs take priority over the keyword arguments in kwargs.

EnsembleKalmanProcesses.Visualize.plot_ϕ_mean_over_time — Function

Visualize.plot_ϕ_mean_over_time(gridposition, ekp, prior, dim_idx; plot_std = false, kwargs...)

Plot the mean constrained parameter of index dim_idx of prior against time on gridposition.

If plot_std = true, then the standard deviation of the constrained parameter of the ensemble is also plotted.

Any keyword arguments is passed to the plotting function which takes in any keyword arguments supported by Makie.Lines and Makie.Band if plot_std = true.

Keyword arguments passed to line_kwargs and band_kwargs are merged with kwargs when possible. The keyword arguments in line_kwargs and band_kwargs take priority over the keyword arguments in kwargs.

Visualize.plot_ϕ_mean_over_time(gridpositions, ekp, prior, name; kwargs...)

Plot the mean constrained parameter belonging to distribution name against time on the iterable gridpositions.

Any keyword arguments are passed to all plots.